requirements.txt
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
asyncpg==0.30.0
autoflake==2.3.1
bcrypt==4.3.0
billiard==4.2.1
black==25.1.0
celery==5.5.1
certifi==2025.1.31
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.12
fastapi-cache2==0.2.2
fastapi-cli==0.0.7
fastapi-versioning==0.10.0
flake8==7.2.0
flower==2.0.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humanize==4.12.2
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jose==1.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kombu==5.5.2
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nodeenv==1.9.1
orjson==3.10.16
packaging==24.2
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.4.8
pycodestyle==2.13.0
pydantic==2.11.2
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pyflakes==3.3.2
Pygments==2.19.1
pyright==1.1.399
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310; platform_system=="Windows"
PyYAML==6.0.2
pyzmq==26.4.0
redis==4.6.0
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.1
time-machine==2.16.0
tornado==6.4.2
traitlets==5.14.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
ujson==5.10.0
uvicorn==0.34.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
WTForms==3.1.2
